backend/app/pyproject.toml
[tool.poetry]
name = "virtual-screening-analyst"
version = "0.1.0"
description = ""
authors = ["Ato Toffah <atotoffah@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
# pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
Requests = "^2.31.0"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
pymongo = "^4.3.3"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
